534 - Refactor profile violation code into package
