+++ generator/src/main/java/com/scottlogic/deg/generator/inputs/profileviolation/ProfileViolator.java
@@ -1,4 +1,4 @@
-package com.scottlogic.deg.generator.inputs;
+package com.scottlogic.deg.generator.inputs.profileviolation;
 
 import com.scottlogic.deg.generator.Profile;
